refactor(instances): share encrypted secret payload crypto
